+++ Source/ImoutoRebirth.Navigator/ImoutoRebirth.Navigator/View/FullScreenPreviewView.xaml.cs
@@ -6,17 +6,13 @@ namespace ImoutoRebirth.Navigator.View;
 
 public partial class FullScreenPreviewView : UserControl
 {
-    public FullScreenPreviewView()
-    {
-        InitializeComponent();
-    }
+    public FullScreenPreviewView() => InitializeComponent();
 
     private void OnPreviewMouseWheel(object sender, MouseWheelEventArgs e)
     {
-        var vm = DataContext as FullScreenPreviewVM; // Замените на ваш тип ViewModel
-        if (vm?.NextPreviewCommand?.CanExecute(e.Delta < 0) == true)
-        {
+        var vm = DataContext as FullScreenPreviewVM;
+        
+        if (vm?.NextPreviewCommand?.CanExecute(e.Delta < 0) == true) 
             vm.NextPreviewCommand.Execute(e.Delta < 0);
-        }
     }
 }