+++ Source/ImoutoRebirth.Navigator/ImoutoRebirth.Navigator/ViewModel/ListEntries/ListEntryType.cs
@@ -6,5 +6,6 @@ internal enum ListEntryType
     Gif,
     Image,
     Ugoira,
-    Png
+    Png,
+    WebP
 }